+++ Sources/Relayer/SocketConnectionHandler/ManualSocketConnectionHandler.swift
@@ -1,4 +1,5 @@
 
+import Starscream
 import Foundation
 
 class ManualSocketConnectionHandler: SocketConnectionHandler {
@@ -15,20 +16,4 @@ class ManualSocketConnectionHandler: SocketConnectionHandler {
     func handleDisconnect(closeCode: URLSessionWebSocketTask.CloseCode) throws {
         socket.disconnect()
     }
-}
-protocol WebSocketConnecting {
-    var isConnected: Bool {get}
-    func connect()
-    func disconnect()
-}
-
-import Starscream
-extension WebSocket: WebSocketConnecting{}
-extension WebSocket: WebSocketProtocol{}
-protocol WebSocketProtocol {
-    var isConnected: Bool {get}
-    var onConnect: (() -> ())? { get set }
-    var onDisconnect: ((Error?) -> ())? { get set }
-    var onText: ((String)->())? { get set }
-    func write(string: String, completion: (() -> ())?)
 }
