+++ Sources/Relayer/SocketConnectionHandler/WebSocketConnecting.swift
@@ -1,0 +1,10 @@
+import Starscream
+import Foundation
+
+extension WebSocket: WebSocketConnecting{}
+
+protocol WebSocketConnecting {
+    var isConnected: Bool {get}
+    func connect()
+    func disconnect()
+}
